monitor_service/requirements.txt
requests
m3u8
python-dotenv
structlog
tenacity
circuitbreaker
httpx